Fix flaky E2E tests and clean up CI output
- Use Playwright's built-in webServer config instead of custom server management
- Add globalSetup for database initialization (runs once before all tests)
- Remove verbose logging that was cluttering CI output
- Always upload test artifacts (not just on failure)
- Increase test timeout to 60s for CI stability
- Make chromium-no-auth project depend on setup to avoid race conditions
