+++ .github/workflows/e2e-tests.yml
@@ -114,17 +114,17 @@ jobs:
           PGPORT: 5432
 
       - name: Upload test results
-        if: failure()
+        if: always()
         uses: actions/upload-artifact@v4
         with:
           name: playwright-report
           path: e2e/playwright-report/
           retention-days: 7
 
-      - name: Upload test screenshots
-        if: failure()
+      - name: Upload test artifacts
+        if: always()
         uses: actions/upload-artifact@v4
         with:
-          name: playwright-screenshots
+          name: playwright-artifacts
           path: e2e/test-results/
           retention-days: 7
